+++ tests/api/Maintenance/MaintenanceTaskServiceTest.cs
@@ -1,5 +1,4 @@
-﻿using BDMS.Maintenance;
-using BDMS.Services;
+﻿using BDMS.Services;
 using Microsoft.Extensions.Logging;
 using Moq;
 using Moq.Protected;
@@ -7,7 +6,7 @@ using System.Net;
 using System.Net.Http.Json;
 using System.Text.RegularExpressions;
 
-namespace BDMS;
+namespace BDMS.Maintenance;
 
 [TestClass]
 public class MaintenanceTaskServiceTest : MaintenanceTaskTestBase
